CR - JudgmentCriteria int to long and other minor changes

+++ saos-import/src/test/java/pl/edu/icm/saos/importer/notapi/common/JudgmentImportProcessReaderTest.java
@@ -55,7 +55,7 @@ public class JudgmentImportProcessReaderTest {
         
         // assert
         
-        List<Integer> internalRJudgmentIds = Whitebox.getInternalState(scjImportProcessReader, "rJudgmentIds");
+        List<Long> internalRJudgmentIds = Whitebox.getInternalState(scjImportProcessReader, "rJudgmentIds");
         
         verify(rawJudgmentRepository).findAllNotProcessedIds(RawSourceScJudgment.class);
         verifyNoMoreInteractions(rawJudgmentRepository);
